+++ wondrous-app/.env.development
@@ -5,4 +5,4 @@ NEXT_PUBLIC_RPC_URL_ARBITRUM=https://arb1.arbitrum.io/rpc/
 NEXT_PUBLIC_RPC_URL_BSC=https://bsc-dataseed.binance.org/
 NEXT_PUBLIC_RPC_URL_BOBA=https://mainnet.boba.network/
 NEXT_PUBLIC_RPC_URL_GOERLI=https://goerli.infura.io/v3/d828b09eb88d4a3fa4299e29b014bd28
-NEXT_PUBLIC_GRAPHQL_SERVER_URL=https://apistaging.wonderapp.co/graphql
+NEXT_PUBLIC_GRAPHQL_SERVER_URL=http://localhost:4000/graphql
