+++ wondrous-app/components/CreateEntity/CreateEntityModal/PodSearch/index.tsx
@@ -1,4 +1,5 @@
-import { useState } from 'react';
+import DropdownSearch from 'components/DropdownSearch';
+import { useMemo, useState } from 'react';
 import {
   PodSearchAutocomplete,
   PodSearchAutocompletePopper,
@@ -20,16 +21,19 @@ import {
 } from './styles';
 
 function PodSearch(props) {
-  const { options, onChange, value, disabled } = props;
-  const [anchorEl, setAnchorEl] = useState(null);
-  const handleClick = (event) => setAnchorEl(anchorEl ? null : event.currentTarget);
-  const handleClickAway = () => setAnchorEl(null);
-  const open = Boolean(anchorEl);
-  const selectedValue = options.find((option) => option.value === value);
+  const { options, onChange, value, disabled, multiple = false } = props;
+  const selectedValue = useMemo(() => {
+    if (multiple) return value;
+    return options.find((option) => option.id === value?.id);
+  }, [value, options, multiple]);
+
   return (
-    <PodSearchClickAway onClickAway={handleClickAway}>
-      <PodSearchWrapper>
-        <PodSearchButton open={open} disabled={!options || disabled} onClick={handleClick}>
+    <DropdownSearch
+      label="Select pod"
+      searchPlaceholder="Search pods"
+      multiple={multiple}
+      anchorComponent={({ onClick, disabled, open }) => (
+        <PodSearchButton open={open} disabled={disabled} onClick={onClick}>
           <PodSearchImageLabelWrapper>
             <PodSearchDefaultImage color={selectedValue?.color ?? `#474747`} />
             <PodSearchLabel>{selectedValue?.label ?? `Select a Pod`}</PodSearchLabel>
@@ -46,52 +50,12 @@ function PodSearch(props) {
             <PodSearchButtonArrowIcon />
           )}
         </PodSearchButton>
-        <PodSearchPopper open={open} anchorEl={anchorEl} placement="bottom-start" disablePortal>
-          <PodSearchAutocomplete
-            value={selectedValue}
-            renderInput={(params) => (
-              <PodSearchInput
-                {...params}
-                ref={params.InputProps.ref}
-                disableUnderline
-                fullWidth
-                autoFocus
-                InputProps={{
-                  ...params.InputProps,
-                  endAdornment: (
-                    <PodSearchInputAdornment position="end">
-                      <PodSearchInputIcon />
-                    </PodSearchInputAdornment>
-                  ),
-                }}
-              />
-            )}
-            disableClearable
-            isOptionEqualToValue={(option, value) => option.value === value?.value}
-            getOptionLabel={(option) => option.label}
-            renderOption={(props, option) => (
-              <PodSearchListItem {...props} isActive={option.value === selectedValue?.value}>
-                <PodSearchDefaultImage color={option?.color ?? '#474747'} />
-                <PodSearchLabel>{option?.label}</PodSearchLabel>
-              </PodSearchListItem>
-            )}
-            PaperComponent={PodSearchPaper}
-            ListboxComponent={PodSearchList}
-            PopperComponent={(params) => <PodSearchAutocompletePopper {...params} />}
-            open={open}
-            options={options}
-            disablePortal
-            onChange={(event, value, reason) => {
-              if (reason === 'selectOption') {
-                onChange(value.value);
-                handleClickAway();
-              }
-            }}
-            blurOnSelect
-          />
-        </PodSearchPopper>
-      </PodSearchWrapper>
-    </PodSearchClickAway>
+      )}
+      value={selectedValue}
+      options={options}
+      disabled={!options || disabled}
+      onChange={onChange}
+    />
   );
 }
