+++ wondrous-app/components/CreateEntity/CreateEntityModal/Templates/TaskTemplatePicker/TemplateEllipsesIcon.tsx
