+++ wondrous-app/graphql/fragments/task.ts
@@ -250,6 +250,44 @@ export const TaskCardFragment = gql`
   ${MediaFragment}
 `;
 
+export const MilestoneCardFragment = gql`
+  fragment MilestoneCardFragment on MilestoneCard {
+    id
+    createdAt
+    createdBy
+    orgId
+    orgProfilePicture
+    orgName
+    orgUsername
+    pods {
+      podId
+      podName
+      podColor
+    }
+    title
+    description
+    priority
+    dueDate
+    status
+    completedAt
+    commentCount
+    additionalData {
+      userMentions
+    }
+    media {
+      ...MediaFragment
+    }
+    privacyLevel
+    label {
+      id
+      name
+      color
+    }
+    points
+  }
+  ${MediaFragment}
+`;
+
 export const HomePageTaskCardFragment = gql`
   fragment HomePageTaskCardFragment on TaskCard {
     id
@@ -554,38 +592,45 @@ export const MilestoneFragment = gql`
     title
     createdAt
     createdBy
+    completedAt
     description
     orgId
-    podId
-    type
+    podIds
     priority
     dueDate
     status
-    links {
-      url
-      displayName
-      type
-    }
     userMentions
-    creator {
-      username
-      profilePicture
+    media {
+      ...MediaFragment
     }
+    orgOrder
+    privacyLevel
+    deletedAt
+    points
+
     org {
       profilePicture
       name
       username
     }
-    pod {
+    pods {
       name
       color
     }
-    orgOrder
-    podOrder
-    assigneeOrder
+    creator {
+      username
+      profilePicture
+    }
+    observers {
+      id
+      username
+      profilePicture
+      firstName
+      lastName
+    }
     commentCount
-    privacyLevel
   }
+  ${MediaFragment}
 `;
 
 export const TaskTemplateFragment = gql`
