+++ wondrous-app/utils/index.ts
@@ -1,3 +1,5 @@
+import { BOARD_TYPE } from './constants';
+
 export const addSearchParamsUrl = (url: string, param: string, paramValue: string) => {
   const newUrl = new URL(url);
   const searchParams = new URLSearchParams(newUrl.searchParams);
@@ -78,4 +80,11 @@ export const getDiscordUrl = () => {
     return 'https://discord.com/api/oauth2/authorize?client_id=917630803314352208&redirect_uri=https%3A%2F%2Fwondrous-app-git-staging-wonderverse.vercel.app%2Fdiscord%2Fcallback&response_type=code&scope=email%20identify';
   }
   return 'https://discord.com/api/oauth2/authorize?client_id=917630803314352208&redirect_uri=http%3A%2F%2Flocalhost%3A3000%2Fdiscord%2Fcallback&response_type=code&scope=email%20identify';
+};
+
+export const getBoardType = ({ orgBoard, podBoard, userBoard }) => {
+  if (orgBoard) return BOARD_TYPE.org;
+  if (podBoard) return BOARD_TYPE.pod;
+  if (userBoard) return BOARD_TYPE.assignee;
+  return BOARD_TYPE.org;
 };
